refactor: 다음 포인트 결정 역할 Point class 로 역할 변경
