+++ src/main/java/laddergame/config/AppConfig.java
@@ -1,7 +1,7 @@
 package laddergame.config;
 
 import laddergame.controller.LadderController;
-import laddergame.domain.ladder.RandomLadderGenerator;
+import laddergame.domain.generator.RandomLadderGenerator;
 import laddergame.service.LadderGame;
 import laddergame.view.InputView;
 import laddergame.view.OutputView;
