+++ src/main/java/laddergame/domain/generator/RandomLadderGenerator.java
@@ -1,8 +1,14 @@
-package laddergame.domain.ladder;
+package laddergame.domain.generator;
 
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Stream;
+import laddergame.domain.ladder.Ladder;
+import laddergame.domain.ladder.LadderGenerator;
+import laddergame.domain.ladder.LadderHeight;
+import laddergame.domain.ladder.Line;
+import laddergame.domain.ladder.LineSize;
+import laddergame.domain.ladder.Point;
 
 public class RandomLadderGenerator implements LadderGenerator {
 
@@ -18,19 +24,11 @@ public class RandomLadderGenerator implements LadderGenerator {
     }
 
     private Line createLine(final LineSize lineSize) {
-        final List<Point> points = Stream.iterate(generatePoint(), this::decideNextPoint)
+        final List<Point> points = Stream.iterate(generatePoint(), point -> point.next(generatePoint()))
                 .limit(lineSize.getLineSize())
                 .toList();
 
         return new Line(points);
-    }
-
-    private Point decideNextPoint(final Point before) {
-        if (before.isExist()) {
-            return Point.EMPTY;
-        }
-
-        return generatePoint();
     }
 
     private Point generatePoint() {
